+++ crates/fluvio-smartmodule-derive/src/generator/filter.rs
@@ -1,6 +1,10 @@
 use quote::quote;
 use proc_macro2::TokenStream;
+
 use crate::SmartModuleFn;
+use crate::util::ident;
+
+use super::transform::generate_transform;
 
 pub fn generate_filter_smartmodule(func: &SmartModuleFn) -> TokenStream {
     let user_fn = &func.name;
@@ -10,85 +14,31 @@ pub fn generate_filter_smartmodule(func: &SmartModuleFn) -> TokenStream {
         super:: #user_fn(&record)
     );
 
-    quote! {
+    generate_transform(
+        ident("filter"),
+        user_code,
+        quote! {
+            for mut record in records.into_iter() {
 
-        #[allow(dead_code)]
-        #user_code
-
-        #[cfg(target_arch = "wasm32")]
-        mod __system {
-
-            use fluvio_smartmodule::dataplane::smartmodule::{
-                SmartModuleInput, SmartModuleTransformErrorStatus,
-                SmartModuleTransformRuntimeError, SmartModuleKind, SmartModuleOutput
-            };
-            use fluvio_smartmodule::dataplane::core::{Encoder, Decoder};
-            use fluvio_smartmodule::dataplane::record::{Record, RecordData};
-
-
-            #[no_mangle]
-            #[allow(clippy::missing_safety_doc)]
-            pub unsafe fn filter(ptr: *mut u8, len: usize, version: i16) -> i32 {
-
-
-                // copy data from wasm memory
-                extern "C" {
-                    fn copy_records(putr: i32, len: i32);
-                }
-
-                let input_data = Vec::from_raw_parts(ptr, len, len);
-                let mut smartmodule_input = SmartModuleInput::default();
-                if let Err(_err) = Decoder::decode(&mut smartmodule_input, &mut std::io::Cursor::new(input_data), version) {
-                    return SmartModuleTransformErrorStatus::DecodingBaseInput as i32;
-                }
-
-                let records_input = smartmodule_input.record_data;
-                let mut records: Vec<Record> = vec![];
-                if let Err(_err) = Decoder::decode(&mut records, &mut std::io::Cursor::new(records_input), version) {
-                    return SmartModuleTransformErrorStatus::DecodingRecords as i32;
-                };
-
-
-                // PROCESSING
-                let mut output = SmartModuleOutput {
-                    successes: Vec::with_capacity(records.len()),
-                    error: None,
-                };
-
-                for mut record in records.into_iter() {
-
-                    let result = #function_call;
-                    match result {
-                        Ok(value) => {
-                            if value {
-                                output.successes.push(record);
-                            }
-                        }
-                        Err(err) => {
-                            let error = SmartModuleTransformRuntimeError::new(
-                                &record,
-                                smartmodule_input.base_offset,
-                                SmartModuleKind::Filter,
-                                err,
-                            );
-                            output.error = Some(error);
-                            break;
+                let result = #function_call;
+                match result {
+                    Ok(value) => {
+                        if value {
+                            output.successes.push(record);
                         }
                     }
+                    Err(err) => {
+                        let error = SmartModuleTransformRuntimeError::new(
+                            &record,
+                            smartmodule_input.base_offset,
+                            SmartModuleKind::Filter,
+                            err,
+                        );
+                        output.error = Some(error);
+                        break;
+                    }
                 }
-
-                // ENCODING
-                let mut out = vec![];
-                if let Err(_) = Encoder::encode(&output, &mut out, version) {
-                    return SmartModuleTransformErrorStatus::EncodingOutput as i32;
-                }
-
-                let out_len = out.len();
-                let ptr = out.as_mut_ptr();
-                std::mem::forget(out);
-                copy_records(ptr as i32, out_len as i32);
-                output.successes.len() as i32
             }
-        }
-    }
+        },
+    )
 }
