+++ crates/fluvio-smartmodule-derive/src/generator/transform.rs
@@ -1,0 +1,67 @@
+use quote::quote;
+use proc_macro2::TokenStream;
+use syn::{ItemFn, Ident};
+
+pub(crate) fn generate_transform(
+    name: Ident,
+    user_code: &ItemFn,
+    transform: TokenStream,
+) -> TokenStream {
+    quote! {
+
+        #[allow(dead_code)]
+        #user_code
+
+        #[cfg(target_arch = "wasm32")]
+        mod __system {
+            #[no_mangle]
+            #[allow(clippy::missing_safety_doc)]
+            pub unsafe fn #name(ptr: *mut u8, len: usize, version: i16) -> i32 {
+                use fluvio_smartmodule::dataplane::smartmodule::{
+                    SmartModuleInput, SmartModuleTransformErrorStatus,
+                    SmartModuleTransformRuntimeError, SmartModuleKind, SmartModuleOutput,
+                };
+                use fluvio_smartmodule::dataplane::core::{Encoder, Decoder};
+                use fluvio_smartmodule::dataplane::record::{Record, RecordData};
+
+                // DECODING
+                extern "C" {
+                    fn copy_records(putr: i32, len: i32);
+                }
+
+                let input_data = Vec::from_raw_parts(ptr, len, len);
+                let mut smartmodule_input = SmartModuleInput::default();
+                if let Err(_err) = Decoder::decode(&mut smartmodule_input, &mut std::io::Cursor::new(input_data), version) {
+                    return SmartModuleTransformErrorStatus::DecodingBaseInput as i32;
+                }
+
+                let records_input = smartmodule_input.record_data;
+                let mut records: Vec<Record> = vec![];
+                if let Err(_err) = Decoder::decode(&mut records, &mut std::io::Cursor::new(records_input), version) {
+                    return SmartModuleTransformErrorStatus::DecodingRecords as i32;
+                };
+
+
+                // PROCESSING
+                let mut output = SmartModuleOutput {
+                    successes: Vec::with_capacity(records.len()),
+                    error: None,
+                };
+
+                #transform
+
+                // ENCODING
+                let mut out = vec![];
+                if let Err(_) = Encoder::encode(&mut output, &mut out, version) {
+                    return SmartModuleTransformErrorStatus::EncodingOutput as i32;
+                }
+
+                let out_len = out.len();
+                let ptr = out.as_mut_ptr();
+                std::mem::forget(out);
+                copy_records(ptr as i32, out_len as i32);
+                output.successes.len() as i32
+            }
+        }
+    }
+}
