+++ makefiles/check.mk
@@ -49,7 +49,7 @@ run-integration-test: build_smartmodules install_rustup_target
 	cargo test  --features wasi -p fluvio-smartengine -- --ignored --test-threads=1
 
 run-smartmodule-test:	build_smartmodules
-	cargo test test_aggregate_map  -p fluvio-smartengine -- --ignored --nocapture
+	cargo test  -p fluvio-smartengine -- --ignored --nocapture
 
 run-k8-test:	install_rustup_target k8-setup build_k8_image
 	cargo test --lib  -p fluvio-sc  -- --ignored --test-threads=1
